Account for selectivity; add other refinements
Kotaro's notes: I just modified the code to create an index of
abundance. The previous code was sampling from the true total abundance
but I forgot to take into account the selectivity. With this code, the
index of abundance reflects a sample from the available biomass
(=biomass catchable with the specific gear).
Sean's notes: I then switched back to having separate input and output files; added various error checks and warnings; fixed some problems that occurred if not all years were present in the data file; and fixed the plotting code.