Update change_m.r
change_m.r now calls ss3 -noest with parameters from the .ctl file after changing the .ctl and .dat files to incorporate an environmental linkage such that a report.sso file is created that contains info on the new environmental link parameter. This new info is then used to update the original ss3.par file to include the environmental link parameter as part of the MGparams and changes the starter file to run with the updated .ctl and .dat files with the parameters from the .par file.