added a feature to remove fishery comp data or survey comp data
The default option is to use the lcomp data.
But for the data richness group, we want to investigate cases where there aren't any lcomp data for either fishery or survey or both. 
I therefore added an option to turn ON or OFF these comp data. Default to TRUE i.e use the lcomp data.