Switch from reading in the new .ctl file to the original
Change suggested by Kotaro:

The reason why I am doing is it as follow:

I propose here to start the optimization (the "init" values) at the true
values for each parameters (excluding the rec devs). Therefore, putting
these "new" values of bias ramp-up and ramp-down periods into the
"em.ctl" file (in the EM folder) that has already the true values for
"init".

The reason is that for a data poor situation, if we used the
"control.ss.new" file, we might be beginning the simulation run from a
bad space once we change the rec-dev time series (because the MLE could
be biased). As opposed to beginning form the true values which will be
on average good.