Update argument documentation for change_f
Update the argument documentation for change_f to explain how to use the function if you have multiple seasons or fisheries. The code worked for these situations but how to implement it with the current arguments was not self-explanatory. The package must be rebuilt for these changes to take effect in within R.