Create sample_ccomp
A first hack at manipulating the sample_agecomp so that
conditional age-at-length data can be sampled.
The function probably contains many errors but needed to be committed.
Input is more complicated than traditional composition data
because there may be different bins for each year for each fleet.
Currently using the same multinomial / dirichlet sampling scheme
but I need to figure out a way to determine the number of samples
based off of the true length comps. This remains to be done.