Fix error in test, remove test for list, add NULL to bin_vector.
Fixed error in a dual test for ages in bin_vector beyond the population
age bins. The test failed when "age" was not included in type.
Removed the test for bin_vector being a list, as if you only add
size at age data there is no need for an input to bin_vector.
For the previous case, bin_vector was set to NULL as a default.