Fix sample_agecomp.R - Cole's fix had brackets in wrong spots.
sample_agecomp.R will only work if the age comp and conditional
age-at-length data are separated prior to the sampling of the age comps.
Then the matrices must be pasted back together. Cole wrote if statements
for the various scenarios of presence / absence of age and conditional
age-at-length. When calculating the number of rows of conditional data
the trailing bracket was in the wrong spot. It is now fixed.