Fix growth_estimation.R so that CV and sigma correct.
Confusion arose between converting from CV to sigma, where sigma is
used in the likelihood and cv refers to the mean size-at-age data.
This fix includes changes to change_e because another column of data,
the mean size-at-age from the samples was needed.
The slope is now calculated from the cv_old and cv_young but
not returned as a parameter value.