Add bin_vector from data if "age" is used but does not have an assigned bin_vector
"age" may not have an assigned bin_vector when bins are not changing.
In the previous case there is now a check that will assign the bin_vector
from the dat file to "age" and or "len" when the sample function is called
but the user has not specified anything. The previous is also true
when the user does not call sample_agecomp or sample_lencomp but is using
another sampling function that relies on the previous sampling functions.