Fix bug in change_e regarding intVal
If intVal was a character (because of scientific notation) then the routine failed.
Now if intVal (which is passed via par_int vector) is a character then
the routine converts it to a numeric value. This is only done after checking to
see if the value is.na() is TRUE or FALSE, such that as.numeric will not give an error
message.