Create checks in mlacomp to check if mwacomps also exist
If so then the matrices are binded together at the end.
The previous will create problems if mwacomps are created for another
sampling function only to be deleted later. Need to make sure
that if the comp data are created for other reasons that they are then
deleted if sample_mwacomp is NULL.