Add mwacomp to the package.
Added arguments for sample_mwacomp to the package in change_bin,
run_ss3sim, and ss3sim_base. Not sure if arguments need to be
added elsewhere. Also the sampling function sample_mwacomp may have
a minor error in it due to the CV for length being used rather than
a cv for weight. I need to talk to Cole and Juan to see where this
should come from. The package runs with all of the new additions.