Fix sample_mlacomp and clean_data to match Fleet or FltSvy column name.
r4ss was inconsistent in their naming structure and did not match
the column name for Fleet in mlacomp data, using Fleet,
rather than Flt/Svy, which is used in other data structures.
The issue was pointed out to @taylori and is now fixed in r4ss.
To make sure that ss3sim is compatible with the all versions of r4ss
I shortened the reference to the fleet column by using mlacomp$Fl.
This was seen as a better solution than hardcoding the column number
(i.e., mlacomp[,3]) or
making the package incompatible with older versions of r4ss.