Fix ss3sim_base assignment to datfile with index and mlacomp.
With the datfile no longer being read in each time sampling is executed,
the same datfile must be used throughout.
An incorrect assignment to ``datfile`` was being used with index
and mlacomp, assigning it within ``with`` rather than outside.
``ss3sim_base`` was changed to fix the error and assign the changes
to ``datfile`` outside of ``with``.