Fix calculate_re.R for situations when column names do not match.
Column names with "_om" and "_em" did not match and so subsetting
the dataframes led to a different number of dimensions for the re calc.
The script now subsets for only those parameter names that are in both
the em_cols and the om_cols rather than just copying over the names.