Fix bug in id_scenarios where it was picking up non-scenario folders
Fix leads to only folders being picked up with two letters or
numbers that are followed by a dash with then the species name.
Prior to the fix it was picking up a folder with any number of
letters that was followed by a dash, such that ss3-test was
thought to be a scenario. Now there can only be two letters,
numbers, or combination or letters and numbers before the dash.