Fixed incorrect warning generation in clean_data
sample_mlacomp uses # if there is no data instead of NULL to
facilitate the use of nrow instead of NROW. I added a check in
clean_data if the mean length-at-age observations are # to
change them to NULL so that if years is also NULL the correct
amount of removed data is listed, which would be 0 lines.
I did not want to change the use of # inside of sample_mlacomp
in case somewhere else in the package it uses nrow instead of NROW
which would lead to an error or something funny.