Change mean_outfile used for external estimation of growth.
Name of mean_outfile non longer needs to contain .csv and if
it contains remove the data used to generate data for external
growth will be removed automatically using clean_data.
This creates an option for the data to remain within the dat file
for estimation, where it was hardcoded before. This will also allow
for future additions where other datafile names can maybe specify
years that should be allowed to remain.