Change how min and max ages for external growth estimation are calculated.
No longer need the dat file for the external estimation of growth,
though the file is needed elsewhere in change_e.
The function is changed to use the min and max ages of fish in the
external data, rather than looking in the model or data used for the EM
to find min and max ages.