Add check to R\clean_data.r for zero rows.
Previously if years were NULL and all data was removed from
a data type the list element would be a data.frame with zero rows
rather than NULL. This would lead to [1] #blah blah being
placed in the .dat file. I added a check at the end so that
if any of the three data.frames in the datfile have NROW == 0
then that element of the list is set to NULL, which works with
r4ss:SS_writedat.