Fixed a bug in change_tv.
The code searches for a number, which may
be arbitrarily found in another line. The new checks make sure the
number is in the correct place and has the word ENV in it if more
than one line is found. I found this bug when working with selectivity
parameters. I did a double check (i.e., checks are now repetitive) to
make sure this does not happen again.