change indexer from q to w, b/c q is bad and check for tabs along with spaces
q is not a good indexer because it is a function within R so it was changed
to w

Checking for tabs, even though users should not have tabs in their files.
Files should be ss_new files which will not have tabs, but the old cod OM
had tabs, and that is what I was testing with.