create the correct number of areas and vessels in forecast file
SS_writeforecast and SS_readforecast do not play nicely with eachother when
there are zero fleets and areas, zero meaning zero values in the forecast
file. I used a workaround to see if there are zero fleets and areas and
then override those values in the read in forecast file.