Change and add input checks for sample_mlacomp.
Combine the input check of Lbin_lo and Lbin_hi in case both exist at
the same time. Also add a check to ensure that the CV_young and CV_old
match. The check is necessary because only CV_young is used, yet it is
assumed that the values are the same otherwise linear interpolation should
be done. Which could be something we add to the package in the future.