Change to empirical sampling if multinomial is used in sample_mlacomp
When sampling from a multinomial for the mlacomp data, one could randomly
obtain more samples of lengths in a given age bin than there were observed
for the agecomp data. Using sample() there is now no way to sample more aged
fish for lengths than the number of observed fish in that bin. This protocol
is mirrored from sample_calcomp and is not needed for the Dirichlet
distribution you are assuming your sample sizes are so large that you would
never oversample a bin. Dirichlet returns proportions, which is why the
> 1 if statement works. If we switch to using a DirichletMultinomial as
suggested by @colemonahhan then the if statement would fail.