Change estimate argument in change_lo_hi
Estimate before was only TRUE or FALSE, now NULL is the default.
Where the default will not change the phase in which parameters are
estimated. A single value of TRUE or FALSE will be repeated for all
parameters, or you can supply a vector the length of parameters.
This facilitates the use of change_lo_hi within standardize_parameters
when you do not want to change how parameters are estimated.