Rework standardize_bounds
Add functionality for when the parameter names are
different because of case or when grep will not pick up a
punctionation mark such as ( in RO.

Add verbose as an argument to stop the output from SS_parlines, which
currently has no other way to stop the output, even though it has its
own verbose argument. I have posted an issue on the r4ss github regarding
this.