Fix index for Nsamp in sample_calcomp
Nsamp was indexed with [[fl]] rather than [[i]]. fl tells you what
the actual fleet is, where i tells you the index of that fleet in the
current loop. Nsamp is structured the same as years, so we want
to index rather than subsample for a given fleet. The error was caught
when fleets = c(2).