Fix check for pars not in both OM and EM
Using the unique call it was subsetting for parameters that were either
in the OM or EM and not pars only present in both.
Now it removes pars not present in both the OM and EM. If no such pars
exist the NROW of the data frame is zero and the next section is skipped.