Fix bug for windows 86 users
The Sys.getenv call on an 86 machine is /i386 or something, which
makes get_bin think you are on a 32 bit machine when you really are not.
Now it checks for a 3 and then secondarily checks for the 86.
If 86 it passes and moves on, if not it assumes you are on a 32 and
will give you the warning. Though we may want it to stop and not
just give a warning.