Add a line to update devtools b4 installing ss3sim
The newest version of devtools is not on CRAN and it fixes a bug that leads to dependent packages not being downloaded properly. You must either run the devtools::install_github("hadley/devtools") or install.packages("ss3sim"), which properly installs dependent packages, prior to installing ss3sim from github. This error was noted in issue #235.