Files by GL to add fleets, sexes, and schooling.
The code is diverse and broad covering many old functions,
and many new functions. She was using cod that was up to date
with all commits prior to April 19, 2016 and KFJ tried to make
sure that all other parts of the package were compatible with
ss3sim code as of 2017-10-09. Functions still need to be documented
using the roxygen package, but I wanted to commit first.
Multiple additional folders by GL also exist, but I have not added
them as of yet. Some will be migrated into the local folder for which
the compsampling paper is being ran. The main goal here was to get
the package up to speed so that it could be pushed to github and
installed using devtools::install_github("ss3sim/ss3sim@sampling").