Specify version 3.24 in r4ss functions
The default for most r4ss functions is version 3.30 SS models, so now have to specify version = "3.24" since this code is for SS 3.24 models.

Note that functions \code{SS_changepars), \code{SS_readstarter}, \code{SS_writestarter}, and \code{SS_writeforecast} do not have the argument \code{version}, so version was not specified for these. They may or may not be compatible with SS 3.24 models.