Add check for correct # of q lines for EM
There was no check that the number of fleets with indices changed in the data file was consistent in the control file for the EM, so a check was added. This unfortunately makes it necessary to read in a control file into the function as well.

For now, the code will stop and generate an error message if the number of fleets in the data and ctl files are not consistent, but in the future it would be better to manipulate the control file so it is consistent with the data file.