development: allow for users to pass a name for q
Before, the name for the added q parameter was pasted together via
a fleet number and some hardwired wording.
ss3sim needed it to match user input from another function and thus
the hardwired wording was causing things to break.
The default of name=NULL will still allow for the previous
hardwired wording to work.