Fix: remove OM q lines if fleet not sampled from ctl file
Remove lines in control file pertaining to a fleet if there is no index of abundance. For now, assume only fleets 2 and 3 can have indices of abundance and both 2 and 3 will be present in the OM. Note that a different function created to remove only these lines from the ctl file without touching the data file.

There is currently no mechanism to add q fleets to the OM, only to remove.