Allow model run to occur in different dir
Although in theory this function the in and out files from this function could be in separate directories, the previous code would not have allowed this. These changes add the capability to put the new files in a separate directory, abeit they all need to be in the same one. In addition, -nohess instead of -noest was used during the model run to allow the creation of a .par file during the run.

This function may need more cleanup still.