Remove unneeded code and add TODO items to change_tv()
Code from 3.24 was commented out previously, but this commit deletes it, as it will no longer be needed given that specification of time varying parameters has change in SS 3.30 and that this function will generate a ss.par file to use.

Also, this function works as is, but has limited functionality, so adds comments with TODO items to improve this function in the future.