badfix: will need to make better eventually
Lots of todos for derived quantities
Need to move these values to scalar and ts I think, but right now
this commit makes things work regarding the derived quantities.
This data frame allows you to compare the OM to estimated confidence
intervals of timeseries and point estimates for management, thus,
determining if the true value falls within the interval.