develop&fix: Remove fleet and survey checks in change_data
Begin allowing model structures other than those with a single
fleet and a single survey. The removal of these checks in the check_data
function is just the first step of many that will be needed.

Fix the call to ensure that all fleets are in area one by using the
number of fleets from the dat file rather than the arbitrary value of 3.