development: remove bias adjustment iterations
Move towards a bias adjustment framework for every iteration
rather than averaged over a set of iterations and used for the
entire scenario.
For scenarios that investigate recruitment, this change towards,
true bias adjustment will provide more realistic results.
Any move towards best practices in assessments is seen as a positive
move by the team. This one comes at the expense of a longer run time
for the models because you need the Hessian.
The changes have yet to be tested but can be ignored by
setting bias_adjust = FALSE.
Future work needs to be done on making the r4ss::SS_fitbiasramp function
more reliable.