close #252 survey and fleet timing is correct in codOM example
Checked the codOM.dat file and the timing is correct for the fleet,
i.e., -1 and for the survey, i.e., >0. Thus, future corrections will
be put on the user to ensure that their fleet timing is correct rather
than check for it in check data. There are too many things to keep track
of and ss3sim cannot check for all SS options. Instead, we will just
check for things that we manipulate.