development:add tests for sample_lcomp, sample_agecomp
These are relatively simple tests to check that invalid inputs are flagged and that values you would expect to be changed are. It would be great to add a check that the sampling is actually being done as expected (seems hard to write a test for this as the function is right now).