documentation: move outfile to its own template depcricate write_file
write_file was used as a logical if the file should be written or not;
now, outfile defaults to NULL that signifies a file should not be written
such that write_file is no longer needed. A few functions still use
write_file, but all of the sampling ones have been converted.