document: calculate_re remove big example, add info, better code
Large example under \donttest was running an entire model, which
was removed because it is too difficult to keep up all of the full
examples for each function. Better to keep full examples contained
to run_ss3sim, or the like.
Added information regarding what is returned.
Implemented use of intersection to decrease lines of code.
Changed NAN and Inf to NA values.