Remove .Rmd README and just have .md
I forget to update the README.md from the .Rmd file. You can create a hook
to remind you, but the hooks are not transferable between computers, so it
adds a level of complication that I don't think is necessary given that the
only code that is actually run is `citation('ss3sim')`.

Remove the R results from the Readme.rmd