fix: remove NCPUEObs from dat_list b/c it is 3.24 legacy
Don't need NCPUEObs, it is a duplicate of NCPUE in the dat_list
SS_writedat doesn't use if for 3.3 b/c there are terminators for the
matrices rather than an external object that says how many rows.
Also, removes a note in the check from CRAN compatability. One step closer.