refactor: #219 use iteration instead of replicate
Had to change a column in the result files that are stored
and change the results functions to generate columns that are
iteration rather than replicate. Chose to deprecate replicate
because it would have involved more code changes that were
argument inputs to do it the other way. Also, I googled the use
of either term and replicate seemed to be used more.