fix(#218): check that model has reached stable catches
Wrote a check in profile_fmsy to ensure that the model is begin ran
long enough to reach stable catches based on the CV of the catches.
Used the half the number of ages that are modelled to determine how
many years of catches from the last year the model should be stable.
This will probably need fixes later given explorations.

Removed the check that the model was ran for at least one year.