fix: recruitment deviations were 0 if not specified
Fix the mistmatch between the number of years in the model and
specified recruitment deviations. Thanks to @colemonnahan for finding
this.
Update some of the documentation.
Allow for users to specify a named vector if they only want to change
a few years to be different than zero.
Better warning messages when users give their own rec devs and values
of NA are in the resulting vector.
Must match the number of years when running through ss3sim_base.