CRAN(1.0.3)- Introduction vs introduction vs ..\doc\introduction
Use introduction b/c it was breaking on other non-windows platforms.
Got an email from CRAN that 1.0.2 had one more bad link and it was one that
I didn't fix, "Introduction.html", where the others I had made
..\doc\introduction.html, then I realized that all the other links were
working and they didn't have ..\doc\... so it was originally the case
of the letter that was wrong.