feature(change_year): allow for EM to have different start year
Issue #254 revealed differences in EM that started with a year of zero
catches and one that didn't. The change_year function looks at the data
to determine when the EM should start.

This function requires a different version of r4ss that is in development.
Only committing the functions right now and not the code changes or tests
that implement them because r4ss will need to be updated first.