feature(#254): change start year of the EM to first year of catches
Model estimates are closer to the OM when the EM doesn't contain years
without catch data, i.e., the burn in period of the OM.
EM now starts with the first year of catches and the early recruitment
period is based on when animals are mature, i.e., generation time from
the start of composition data. For cod, the default is to start in year 6
for recruitment.