feature: allow bad scenario names, skip empty report and forecast
(1) users might have scenarios that are different than the traditional
names that they still want read in, this new feature just allows ss3sim
to check that the folders exist in the directory and they don't have to
be true ss3sim scenario names, this will help with ss3sim being more
applicable to other simulation frameworks.

(2) remove the warning message if the csv file is not available but
ss3sim is trying to read it in; allowing for empty scenario folders
to not cause errors and facilitate reading in results mid stream of
a simulation.

(3) skip a scenario if the report file is empty but has been written
to the disk while the m odel is still running.

(4) don't read in the forecasts for a scenario that doesn't have a
proper forecast report file, to facilitate reading in the results
mid stream.