fix: removal and addition of q based on data
This fix has changes for both the OM and the EM as well
as a testing file that checks some addition and removal of q
values but does not run ss3sim. When adding a q parameter there
is currently no way to add extra_se, just a simple ln_q parameter
with no float set to a value of zero or (exp[0], which equal 1).
The changes do the following:

* Remove remove_q
* remove remove_CPUE
* remove add_CPUE
* add check_q
* add change_q, which will both add and remove q parameters.