fix(ss3sim_base): remove loop for multiple scenarios
ss3sim_base is only meant to run a single scenario yet there was a loop
for multiple scenarios. run_ss3sim(), which calls ss3sim_base(), parses
out the case files or parameters for each scenario and feeds them to
ss3sim_base() as a list one scenario at a time.

A stopifnot call was put in to check that scenarios only contains one
scenario. Might change the input argument name later to increase clarity
but it didn't seem worth it at the momement.