close #273 use dir instead of scenarios and iterations
Decrease complexity regarding file paths and passing multiple
arguments when the scenarios and iterations args were not
being used in a vector capacity. Instead, it is much cleaner
to pass a single file path. This will allow some functions
to be used outside of ss3sim rather than just for a structure
that follows scenario/iteration/em.