refactor(get_results_all): reduce seq of if statements to 1
write.csv was being called two times for each file because there
were two if statements. I added the inverse of overwrite_files to
the first if statement to get rid of the else. Now if someone wants
to change how the files are written in the future, such as was done
in this branch they wouldn't have to find write.csv two times for
each file.