fix(change_q): remove stop if args are NULL
Removed the stop() call when string_add and string_remove are NULL.
next was already in place in the for loop to exit out of the loop if
nothing needed to be changed. After next, then the next two calls will
be FALSE and the read in file is written back to the disk with the correct
name and returned invisibly.