feat(setup_scenarios): allow for NA in all fleet columns
NAs no longer need to be in Nsamp, can be in any column, which are
transformed to lists, such as cf.years.1.
Doesn't need NA in every column either to remove a fleet, just one will do.
Created helper functions to make it easier to see where you are in the
list structure rather than having so many for loops.

Works to remove fleets in the middle such as fleet 2 when you want to
keep fleet 1 and 3, thanks to using the original indexing put forth by
@k-doering-noaa.

Added more tests while I was developing and decided to leave them b/c
they are quick.