fix(test-setup_scenarios): need to test on comps not input list
Test needs to be using the sampling function and not on equivalent data frames
because I operate under the assumption that the function will handle the missing
length of lists. For example, sample_comp handles the mis-match in list lengths.

Second, you can only do this for all but one entry within a category. At least
one fleet specific value needs to be present otherwise the functions don't
know how long to make the lists. So, I broke the test into two separate tests.