feat(equilibrium catch): Implement equilibrium catch if F year == -999
Prior, equilibrium catches were always set to 0. With this added feature,
one can set equilibrium catches using the year of -999 in your data frame
for F values to the OM. This implements an F parameter for equilibrium
catches in the OM.

change_f
* remove deprecated fisheries input argument and the warning about it being
deprecated
* remove control file input and output names in favour of using an control list
as read in by SS_readctl to limit the number of times that the file is read and
written to the disk; this will be carried forward into other functions as well.
* make down-stream functions work with these changes within ss3sim

change_catch
* use both ctl and data file lists rather than input arguments to reduce the
chance for potential mistakes when creating the F data frame
* use proper parameters rather than just an F over years
* remove catches that are not needed, instead just use catches for year,seas,...
combinations that have a positive F value to make the data file smaller

Close #302