fix(verify_input): use $ for end of file search with grep
When running Chang's model, where I had run ss prior on the OM
prior to trying to use that same folder for a simulation, I
found that the starter.ss_new and forecast.ss_new files
lead to a matrix rather than a single row of input files to copy.
This leads to a cryptic message about lengths not matching.
Now, the $ terminates the grep telling it that .ss are the last
characters of the string to search for.