feat(vbgf): remove external estimates of growth
bbmle was an R package needed to externally estimate growth
from samples created within ss3sim. I would love to add this feature
back in the future but right now it was not being used by the user
community and bbmle was often not installed and led to increased
fatigue in installing ss3sim. I think in the future we could find
alternative ways to estimate growth and supply input values to
ss3sim.