doc: use _PACKAGE renamed ss3sim-package.R
[roxygen2 documentation](https://cran.r-project.org/web/packages/roxygen2/vignettes/rd.html)
suggests using _PACKAGE to denote the .R file for
package documentation. It was also suggested that this file
be named {package}-package.R so users can easily see that
the file documents the package. I followed the style of
ggplot2 for these changes because some tidyverse packages
have yet to update their files to this new structure.