fix(person): Implements one family for Giancarlo
pkgdown:::build_citation_authors() had two
authors for Giancarlo because
I was splitting his name into a vector.
Now the name is combined with a space
and there is no longer an error about
not being able to check the name.
Hopefully this will fix pkgdown
from not being able to build
the website using GitHub actions.