fix: Do NOT turn on empirical
If sampling empirical weight at age, do NOT
automatically turn on empirical weight at age in the EM.
Users can instead pass two different EMs, one with it turned
on and one with it turned off to use the data or not.
This allows for separation between the OM and EM because
you might just want to generate the data but not use it.

Part of #300