Updates data structures to use SS3.30.19.1 output
Adds code to cod.R to copy ss.exe and run the OM,
move the new dat file to the EM and run it as well.
Files are saved, read in using r4ss, and saved as data objects.
Multiple tests and examples were changed to accommodate
new naming structure of natural mortality, selectivity, etc.