fix(copy_ss3models): Removes vectorization
Removes the two loops inside `copy_ss3models()` and changes the
arguments scenario and iteration to be singular rather than plural.

I checked into using `r4ss::copy_SS_inputs()` instead but it was slower
and I would of had to verify the model afterwards as well. For right
now I will just keep this function.

Fixes the output to match the input in that prior the input allowed for
the creation of multiple directories but the return boolean was only
for the last directory that was created. Now, there is only one directory
created so the input/output match.

Close #276