fix(get_bin): Removes `shQuote()` for `system2()`
As noted by @iantaylor-NOAA the use of `system2()`
and maybe other things negates the need for wrapping
file paths with quotes with `shQuote()`. Thus, the actual
problem in https://github.com/r4ss/r4ss/issues/747 was
not with {r4ss} but rather with {ss3sim}. Quoting is
now removed and I checked it with
`"C:/Program Files/R/R-devel/library/ss3sim/bin/Windows64/ss3.exe"`
and it works.