fix (sample_index()): corrects duplication of seas
when there is more than one fleet

seas needs to be a list but most of the time it is ignored because we do not
really deal with season. The example that was emailed to me had two surveys,
where we often only deal with a single survey. Thus, seas = list(1) was being
repeated given the number of fleets. This should have been seas = list(1,1)
rather than list(1) which was repeated incorrectly using rep(seas, Nfleets).
The bad code is removed and the default value of 1 will be repeated for all
fleets automatically in the checks that are in standardize_sampling_args().

enhance error checking of inputs

* remove Nfleets == 0 from the checks because it is now its own check and if
  zero, then the dat_list is returned early because no sampling should take
  place.
* use cli::cli_abort() instead of stop to allow for more detailed error
  messages and align with tidyverse.
* group all input changes/set up at the top
* group all checks after the set up (before they were mixed)
* remove checks because many are checked for in standardize_sampling_args().
* fix two issues of "no visible binding for global variable" with
  * dist and
  * index